// frontend/src/lib/schema.ts
export type Id = number;
export type Text = string | (string | PlainLike | Pre | TextLink | MentionName | CustomEmoji)[];
// ...
export type TextEntities = (PlainLike | Pre | TextLink | MentionName | CustomEmoji)[];
export type Type5 = "message";
export type FromId = string;
export type From = string;
export type Date = string;
export type DateUnixtime = number;
export type Edited = string | null;
export type EditedUnixtime = number | null;
export type ForwardedFrom = string | null;
export type ReplyToMessageId = number | null;
export type ViaBot = string | null;
export type File = string | null;
export type MimeType = string | null;
export type Photo = string | null;
export type Width = number | null;
export type Height = number | null;
export type StickerEmoji = string | null;
export type Thumbnail = string | null;
export type MediaType =
  | ("sticker" | "animation" | "video_file" | "audio_file" | "video_message" | "voice_message")
  | null;
export type DurationSeconds = number | null;
export type PhoneNumber = string;
export type FirstName = string;
export type LastName = string;
export type ContactVcard = string;
export type Latitude = number;
export type Longitude = number;
export type LiveLocationPeriodSeconds = number | null;
export type InlineBotButtons = BotButton[][] | null;
export type Type6 = "url";
export type Text6 = string;
export type Data = string;
export type Id1 = number;
export type Text7 = string | (string | PlainLike | Pre | TextLink | MentionName | CustomEmoji)[];
export type TextEntities1 = (PlainLike | Pre | TextLink | MentionName | CustomEmoji)[];
export type Type7 = "service";
export type Actor = string;
export type ActorId = string;
export type Date1 = string;
export type DateUnixtime1 = number;
export type Action = "phone_call";
export type DurationSeconds1 = number | null;
export type DiscardReason = ("hangup" | "missed" | "busy" | "disconnect") | null;
export type Id2 = number;
export type Text8 = string | (string | PlainLike | Pre | TextLink | MentionName | CustomEmoji)[];
export type TextEntities2 = (PlainLike | Pre | TextLink | MentionName | CustomEmoji)[];
export type Type8 = "service";
export type Actor1 = string;
export type ActorId1 = string;
export type Date2 = string;
export type DateUnixtime2 = number;
export type Action1 = "edit_chat_theme";
export type Emoticon = string;
// ...
export type Id4 = number;
export type Text10 = string | (string | PlainLike | Pre | TextLink | MentionName | CustomEmoji)[];
export type TextEntities4 = (PlainLike | Pre | TextLink | MentionName | CustomEmoji)[];
export type Type10 = "service";
export type Actor3 = string;
export type ActorId3 = string;
export type Date4 = string;
export type DateUnixtime4 = number;
export type Action3 = "migrate_to_supergroup";
export type Id5 = number;
export type Text11 = string | (string | PlainLike | Pre | TextLink | MentionName | CustomEmoji)[];
export type TextEntities5 = (PlainLike | Pre | TextLink | MentionName | CustomEmoji)[];
export type Type11 = "service";
export type Actor4 = string;
export type ActorId4 = string;
export type Date5 = string;
export type DateUnixtime5 = number;
export type Action4 = "migrate_from_group";
export type Title1 = string;
export type Id6 = number;
export type Text12 = string | (string | PlainLike | Pre | TextLink | MentionName | CustomEmoji)[];
export type TextEntities6 = (PlainLike | Pre | TextLink | MentionName | CustomEmoji)[];
export type Type12 = "service";
export type Actor5 = string;
export type ActorId5 = string;
export type Date6 = string;
export type DateUnixtime6 = number;
export type Action5 = "invite_members";
export type Members1 = string[];
export type Id7 = number;
export type Text13 = string | (string | PlainLike | Pre | TextLink | MentionName | CustomEmoji)[];
export type TextEntities7 = (PlainLike | Pre | TextLink | MentionName | CustomEmoji)[];
export type Type13 = "service";
export type Actor6 = string;
export type ActorId6 = string;
export type Date7 = string;
export type DateUnixtime7 = number;
export type Action6 = "edit_group_photo";
export type Photo1 = string;
export type Width1 = number;
export type Height1 = number;

export interface Message {
  id: Id;
  text: Text;
  text_entities: TextEntities;
  type: Type5;
  from_id: FromId;
  from: From;
  date: Date;
  date_unixtime: DateUnixtime;
  edited?: Edited;
  edited_unixtime?: EditedUnixtime;
  forwarded_from?: ForwardedFrom;
  reply_to_message_id?: ReplyToMessageId;
  via_bot?: ViaBot;
  file?: File;
  mime_type?: MimeType;
  photo?: Photo;
  width?: Width;
  height?: Height;
  sticker_emoji?: StickerEmoji;
  thumbnail?: Thumbnail;
  media_type?: MediaType;
  duration_seconds?: DurationSeconds;
  contact_information?: Contact | null;
  contact_vcard?: ContactVcard;
  location_information?: Location | null;
  live_location_period_seconds?: LiveLocationPeriodSeconds;
  inline_bot_buttons?: InlineBotButtons;
}
export interface PlainLike {
  type: Type;
  text: Text1;
}
export interface Pre {
  type: Type1;
  text: Text2;
  language: Language;
}
export interface TextLink {
  type: Type2;
  text: Text3;
  href: Href;
}
export interface MentionName {
  type: Type3;
  text: Text4;
  user_id: UserId;
}
export interface CustomEmoji {
  type: Type4;
  text: Text5;
  document_id: DocumentId;
}
export interface Contact {
  phone_number: PhoneNumber;
  first_name: FirstName;
  last_name: LastName;
}
export interface Location {
  latitude: Latitude;
  longitude: Longitude;
}
export interface BotButton {
  type: Type6;
  text: Text6;
  data: Data;
}
export interface PhoneCall {
  id: Id1;
  text: Text7;
  text_entities: TextEntities1;
  type: Type7;
  actor: Actor;
  actor_id: ActorId;
  date: Date1;
  date_unixtime: DateUnixtime1;
  action: Action;
  duration_seconds?: DurationSeconds1;
  discard_reason?: DiscardReason;
}
export interface EditChatTheme {
  id: Id2;
  text: Text8;
  text_entities: TextEntities2;
  type: Type8;
  actor: Actor1;
  actor_id: ActorId1;
  date: Date2;
  date_unixtime: DateUnixtime2;
  action: Action1;
  emoticon: Emoticon;
}
export interface CreateGroup {
  id: Id3;
  text: Text9;
  text_entities: TextEntities3;
  type: Type9;
  actor: Actor2;
  actor_id: ActorId2;
  date: Date3;
  date_unixtime: DateUnixtime3;
  action: Action2;
  title: Title;
  members: Members;
}
export interface MigrateToSupergroup {
  id: Id4;
  text: Text10;
  text_entities: TextEntities4;
  type: Type10;
  actor: Actor3;
  actor_id: ActorId3;
  date: Date4;
  date_unixtime: DateUnixtime4;
  action: Action3;
}
export interface MigrateFromGroup {
  id: Id5;
  text: Text11;
  text_entities: TextEntities5;
  type: Type11;
  actor: Actor4;
  actor_id: ActorId4;
  date: Date5;
  date_unixtime: DateUnixtime5;
  action: Action4;
  title: Title1;
}
export interface InviteMembers {
  id: Id6;
  text: Text12;
  text_entities: TextEntities6;
  type: Type12;
  actor: Actor5;
  actor_id: ActorId5;
  date: Date6;
  date_unixtime: DateUnixtime6;
  action: Action5;
  members: Members1;
}
export interface EditGroupPhoto {
  id: Id7;
  text: Text13;
  text_entities: TextEntities7;
  type: Type13;
  actor: Actor6;
  actor_id: ActorId6;
  date: Date7;
  date_unixtime: DateUnixtime7;
  action: Action6;
  photo: Photo1;
  width: Width1;
  height: Height1;
}